+++ API Pessoa.sln
@@ -3,7 +3,25 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.7.34221.43
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "API Pessoa", "API Pessoa.csproj", "{DFB38CD9-047D-46EC-A6AF-58A7F2784D41}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "API Pessoa", "API Pessoa.csproj", "{DFB38CD9-047D-46EC-A6AF-58A7F2784D41}"
+	ProjectSection(ProjectDependencies) = postProject
+		{C8D564E1-84FA-43B6-9749-C94280D92D9D} = {C8D564E1-84FA-43B6-9749-C94280D92D9D}
+	EndProjectSection
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "0-UI", "0-UI", "{47625934-FF46-4FD3-BB3B-88D1890324D1}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "1-CORE", "1-CORE", "{90589948-0110-41E6-8509-211FAB967133}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "2-INFRA", "2-INFRA", "{30F4EBE2-C262-4E28-9FB6-55D1D062914B}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Core", "..\Core\Core.csproj", "{E005061B-DDE4-4B2B-AD75-1A6D7AFA2422}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Data", "..\Data\Data.csproj", "{C8D564E1-84FA-43B6-9749-C94280D92D9D}"
+	ProjectSection(ProjectDependencies) = postProject
+		{E005061B-DDE4-4B2B-AD75-1A6D7AFA2422} = {E005061B-DDE4-4B2B-AD75-1A6D7AFA2422}
+	EndProjectSection
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "2.1-DATA", "2.1-DATA", "{A95C251F-BC9A-435A-B097-5F8030AD9D47}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,9 +33,23 @@ Global
 		{DFB38CD9-047D-46EC-A6AF-58A7F2784D41}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DFB38CD9-047D-46EC-A6AF-58A7F2784D41}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{DFB38CD9-047D-46EC-A6AF-58A7F2784D41}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E005061B-DDE4-4B2B-AD75-1A6D7AFA2422}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E005061B-DDE4-4B2B-AD75-1A6D7AFA2422}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E005061B-DDE4-4B2B-AD75-1A6D7AFA2422}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E005061B-DDE4-4B2B-AD75-1A6D7AFA2422}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C8D564E1-84FA-43B6-9749-C94280D92D9D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C8D564E1-84FA-43B6-9749-C94280D92D9D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C8D564E1-84FA-43B6-9749-C94280D92D9D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C8D564E1-84FA-43B6-9749-C94280D92D9D}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{DFB38CD9-047D-46EC-A6AF-58A7F2784D41} = {47625934-FF46-4FD3-BB3B-88D1890324D1}
+		{E005061B-DDE4-4B2B-AD75-1A6D7AFA2422} = {90589948-0110-41E6-8509-211FAB967133}
+		{C8D564E1-84FA-43B6-9749-C94280D92D9D} = {A95C251F-BC9A-435A-B097-5F8030AD9D47}
+		{A95C251F-BC9A-435A-B097-5F8030AD9D47} = {30F4EBE2-C262-4E28-9FB6-55D1D062914B}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {8B9A4740-CD5D-4608-BE76-3DADBDEF0B26}
