+++ Controllers/PessoaController.cs
@@ -1,23 +1,22 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
-using API_Pessoa.cad_Pessoa.Entities;
-using API_Pessoa.cad_Pessoa.Persistence;
-using API_Pessoa.cad_Pessoa.Models;
-using API_Pessoa.cad_Pessoa.Mappers;
+using API_Pessoa.Core.Entities;
+using API_Pessoa.Data.Context;
+using API_Pessoa.Core.DTOs;
 using System.Reflection.Metadata.Ecma335;
 using AutoMapper;
 using System.Collections;
 
-namespace API_Pessoa.cad_Pessoa.Controllers
+namespace API_Pessoa.Controllers
 {
     [Route("api")]
     [ApiController]
     public class PessoaController : ControllerBase
     {
-        private readonly PessoaContext _context;
+        private readonly DBContext _context;
         private readonly IMapper _mapper;
 
-        public PessoaController(PessoaContext context, IMapper mapper)
+        public PessoaController(DBContext context, IMapper mapper)
         {
             _context = context;
             _mapper = mapper;
@@ -26,14 +25,14 @@ namespace API_Pessoa.cad_Pessoa.Controllers
         [HttpGet("Pessoas")]
         public List<Pessoa> Pessoas()
         {
-            var Pessoas = _context.Pessoas.ToList();
+            var Pessoas = _context.set<Pessoa>().ToList();
 
             return Pessoas;
         }
         [HttpGet("Pessoas/{id}")]
         public async Task<IActionResult> Pessoa(int id)
         {
-            var pessoaPorID = await _context.Pessoas.FirstOrDefaultAsync(x => x.IdPessoa == id);
+            var pessoaPorID = await _context.set<Pessoa>().FirstOrDefaultAsync(x => x.IdPessoa == id);
 
             if (pessoaPorID == null)
                 return NotFound(new { Error = "Pessoa não encontrada." });
@@ -54,17 +53,17 @@ namespace API_Pessoa.cad_Pessoa.Controllers
             if (pessoaInput.IdPais == 0)
             {
                 pessoaInput.IdPais = null;
-            } 
-            if (pessoaInput.IdUnidade_Federativa == 0) 
+            }
+            if (pessoaInput.IdUnidade_Federativa == 0)
             {
                 pessoaInput.IdUnidade_Federativa = null;
-            } 
+            }
             if (pessoaInput.IdMunicipio == 0)
             {
                 pessoaInput.IdMunicipio = null;
             }
 
-            _context.Pessoas.Add(pessoaInput);
+            _context.set<Pessoa>().Add(pessoaInput);
             _context.SaveChanges();
 
             return Ok(pessoaInput);
@@ -84,7 +83,7 @@ namespace API_Pessoa.cad_Pessoa.Controllers
 
         [HttpGet("PegaHoraBrasilia")]
         public DateTime PegaHoraBrasilia() => TimeZoneInfo.ConvertTime(DateTime.Now, TimeZoneInfo.FindSystemTimeZoneById("E. South America Standard Time"));
-        
+
         [HttpGet("Pessoa/novoCodigo")]
         public string novoCodigo()
         {
@@ -126,6 +125,21 @@ namespace API_Pessoa.cad_Pessoa.Controllers
         {
             var test = Pessoas();
             return Ok(test);
+        }
+        /* TESTES DE FILTRAGEM */
+
+        [HttpGet("Pessoas/{nome}/{idPais}")]
+        public IActionResult Filtros(string nome, int idPais)
+        {
+            IQueryable<Pessoa> query = _context.Pessoas;
+
+            if (nome != null)
+                query = query.Where(x => x.Nome == nome);
+            if (idPais != 0)
+                query = query.Where(x => x.IdPais == idPais);
+
+
+            return Ok(query.ToList());
         }
     }
 }
