+++ appsettings.json
@@ -7,6 +7,7 @@
   },
   "AllowedHosts": "*",
   "ConnectionStrings": {
-    "Trabalho": "data source=DESKTOP-H1ES66J;initial catalog=HeadTESTE;trusted_connection=true;trustServerCertificate=true;"
+    "Trabalho": "data source=DESKTOP-H1ES66J;initial catalog=HeadTESTE;trusted_connection=true;trustServerCertificate=true;",
+    "Casa": "data source=LAPTOP-IAL38H7O;initial catalog=HeadTEST;trusted_connection=true;trustServerCertificate=true;"
   }
 }
