--- cad_Cliente/Objt/Entry.cs
@@ -1,6 +1,0 @@
-﻿namespace API_Pessoa.cad_Cliente.Objt
-{
-    public class Entry
-    {
-    }
-}
